Declare constant load/store displacements as long.
  There may be cases where code will handle buffers larger than 2GB.

+++ lightning/i386/core-64.h
@@ -1615,7 +1615,7 @@ jit_ldi_c(jit_gpr_t r0, void *i0)
 
 #define jit_ldxi_c(r0, r1, i0)		jit_ldxi_c(r0, r1, i0)
 __jit_inline void
-jit_ldxi_c(jit_gpr_t r0, jit_gpr_t r1, int i0)
+jit_ldxi_c(jit_gpr_t r0, jit_gpr_t r1, long i0)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVSBQmr(i0, r1, 0,  0, r0);
@@ -1639,7 +1639,7 @@ jit_ldi_uc(jit_gpr_t r0, void *i0)
 
 #define jit_ldxi_uc(r0, r1, i0)		jit_ldxi_uc(r0, r1, i0)
 __jit_inline void
-jit_ldxi_uc(jit_gpr_t r0, jit_gpr_t r1, int i0)
+jit_ldxi_uc(jit_gpr_t r0, jit_gpr_t r1, long i0)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVZBLmr(i0, r1, 0, 0, r0);
@@ -1677,7 +1677,7 @@ jit_stxr_c(jit_gpr_t r0, jit_gpr_t r1, jit_gpr_t r2)
 
 #define jit_stxi_c(i0, r0, r1)		jit_stxi_c(i0, r0, r1)
 __jit_inline void
-jit_stxi_c(int i0, jit_gpr_t r0, jit_gpr_t r1)
+jit_stxi_c(long i0, jit_gpr_t r0, jit_gpr_t r1)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVBrm(r1, i0, r0, 0, 0);
@@ -1701,7 +1701,7 @@ jit_ldi_s(jit_gpr_t r0, void *i0)
 
 #define jit_ldxi_s(r0, r1, i0)		jit_ldxi_s(r0, r1, i0)
 __jit_inline void
-jit_ldxi_s(jit_gpr_t r0, jit_gpr_t r1, int i0)
+jit_ldxi_s(jit_gpr_t r0, jit_gpr_t r1, long i0)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVSWQmr(i0, r1, 0, 0, r0);
@@ -1725,7 +1725,7 @@ jit_ldi_us(jit_gpr_t r0, void *i0)
 
 #define jit_ldxi_us(r0, r1, i0)		jit_ldxi_us(r0, r1, i0)
 __jit_inline void
-jit_ldxi_us(jit_gpr_t r0, jit_gpr_t r1, int i0)
+jit_ldxi_us(jit_gpr_t r0, jit_gpr_t r1, long i0)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVZWLmr(i0, r1, 0, 0, r0);
@@ -1749,7 +1749,7 @@ jit_sti_s(void *i0, jit_gpr_t r0)
 
 #define jit_stxi_s(i0, r0, r1)		jit_stxi_s(i0, r0, r1)
 __jit_inline void
-jit_stxi_s(int i0, jit_gpr_t r0, jit_gpr_t r1)
+jit_stxi_s(long i0, jit_gpr_t r0, jit_gpr_t r1)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVWrm(r1, i0, r0, 0, 0);
@@ -1787,7 +1787,7 @@ jit_ldxr_i(jit_gpr_t r0, jit_gpr_t r1, jit_gpr_t r2)
 
 #define jit_ldxi_i(r0, r1, i0)		jit_ldxi_i(r0, r1, i0)
 __jit_inline void
-jit_ldxi_i(jit_gpr_t r0, jit_gpr_t r1, int i0)
+jit_ldxi_i(jit_gpr_t r0, jit_gpr_t r1, long i0)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVSLQmr(i0, r1, 0, 0, r0);
@@ -1825,7 +1825,7 @@ jit_ldxr_ui(jit_gpr_t r0, jit_gpr_t r1, jit_gpr_t r2)
 
 #define jit_ldxi_ui(r0, r1, i0)		jit_ldxi_ui(r0, r1, i0)
 __jit_inline void
-jit_ldxi_ui(jit_gpr_t r0, jit_gpr_t r1, int i0)
+jit_ldxi_ui(jit_gpr_t r0, jit_gpr_t r1, long i0)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVLmr(i0, r1, 0, 0, r0);
@@ -1849,7 +1849,7 @@ jit_sti_i(void *i0, jit_gpr_t r0)
 
 #define jit_stxi_i(i0, r0, r1)		jit_stxi_i(i0, r0, r1)
 __jit_inline void
-jit_stxi_i(int i0, jit_gpr_t r0, jit_gpr_t r1)
+jit_stxi_i(long i0, jit_gpr_t r0, jit_gpr_t r1)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVLrm(r1, i0, r0, 0, 0);
@@ -1889,7 +1889,7 @@ jit_ldxr_l(jit_gpr_t r0, jit_gpr_t r1, jit_gpr_t r2)
 #define jit_ldxi_ul(r0, r1, i0)		jit_ldxi_l(r0, r1, i0)
 #define jit_ldxi_p(r0, r1, i0)		jit_ldxi_l(r0, r1, i0)
 __jit_inline void
-jit_ldxi_l(jit_gpr_t r0, jit_gpr_t r1, int i0)
+jit_ldxi_l(jit_gpr_t r0, jit_gpr_t r1, long i0)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVQmr(i0, r1, 0, 0, r0);
@@ -1927,7 +1927,7 @@ jit_stxr_l(jit_gpr_t r0, jit_gpr_t r1, jit_gpr_t r2)
 
 #define jit_stxi_l(i0, r0, r1)		jit_stxi_l(i0, r0, r1)
 __jit_inline void
-jit_stxi_l(int i0, jit_gpr_t r0, jit_gpr_t r1)
+jit_stxi_l(long i0, jit_gpr_t r0, jit_gpr_t r1)
 {
     if (jit_can_sign_extend_int_p(i0))
 	MOVQrm(r1, i0, r0, 0, 0);
